read gold cost option as int32
removes redundant int32 casts

no need to store gold cost as uint32; gold cost max is 214748g

+++ src/RacialTraitSwap.cpp
@@ -446,7 +446,7 @@ public:
 
     bool OnGossipSelect(Player* player, Creature* creature, uint32 /*menuId*/, uint32 gossipListId) override
     {
-        int32 const RTS1 = sConfigMgr->GetOption<uint32>("Racial.Traits.Swap.Gold", 100);
+        int32 const RTS1 = sConfigMgr->GetOption<int32>("Racial.Traits.Swap.Gold", 100);
 
         char const* localizedExit = {};
         char const* localizedBack = {};
@@ -1219,9 +1219,9 @@ public:
 
             case 112: // Blood Elf
                 CloseGossipMenuFor(player);
-                if (!player->HasEnoughMoney(int32(RTS1 * GOLD))) // gold check
+                if (!player->HasEnoughMoney(RTS1 * GOLD)) // gold check
                     return true;
-                player->ModifyMoney(int32(-RTS1 * GOLD)); // Deducting the money if check passes
+                player->ModifyMoney(-RTS1 * GOLD); // Deducting the money if check passes
                 player->CastSpell(player, 47292); // Level up visual effect to let you know the transaction did occure and money is deducted
                 RemoveAllRacials(player); // Remove all racials
                 player->learnSpell(28877, false); // Arcane Affinity
@@ -1261,9 +1261,9 @@ public:
 
             case 212: // Draenei
                 CloseGossipMenuFor(player);
-                if (!player->HasEnoughMoney(int32(RTS1 * GOLD))) // gold check
+                if (!player->HasEnoughMoney(RTS1 * GOLD)) // gold check
                     return true;
-                player->ModifyMoney(int32(-RTS1 * GOLD)); // Deducting the money if check passes
+                player->ModifyMoney(-RTS1 * GOLD); // Deducting the money if check passes
                 player->CastSpell(player, 47292); // Level up visual effect to let you know the transaction did occure and money is deducted
                 RemoveAllRacials(player); // Remove all racials
                 player->learnSpell(28875, false); // Gem Cutting
@@ -1333,9 +1333,9 @@ public:
 
             case 31: // Dwarves
                 CloseGossipMenuFor(player);
-                if (!player->HasEnoughMoney(int32(RTS1 * GOLD))) // gold check
+                if (!player->HasEnoughMoney(RTS1 * GOLD)) // gold check
                     return true;
-                player->ModifyMoney(int32(-RTS1 * GOLD)); // Deducting the money if check passes
+                player->ModifyMoney(-RTS1 * GOLD); // Deducting the money if check passes
                 player->CastSpell(player, 47292); // Level up visual effect to let you know the transaction did occure and money is deducted
                 RemoveAllRacials(player); // Remove all racials
                 player->learnSpell(2481, false); // Find Treasure
@@ -1354,9 +1354,9 @@ public:
 
             case 41: // Gnome
                 CloseGossipMenuFor(player);
-                if (!player->HasEnoughMoney(int32(RTS1 * GOLD))) // gold check
+                if (!player->HasEnoughMoney(RTS1 * GOLD)) // gold check
                     return true;
-                player->ModifyMoney(int32(-RTS1 * GOLD)); // Deducting the money if check passes
+                player->ModifyMoney(-RTS1 * GOLD); // Deducting the money if check passes
                 player->CastSpell(player, 47292); // Level up visual effect to let you know the transaction did occur and money is deducted
                 RemoveAllRacials(player); // Remove all racials
                 player->learnSpell(20592, false); // Arcane Resistance
@@ -1374,9 +1374,9 @@ public:
 
             case 51: // Human
                 CloseGossipMenuFor(player);
-                if (!player->HasEnoughMoney(int32(RTS1 * GOLD))) // gold check
+                if (!player->HasEnoughMoney(RTS1 * GOLD)) // gold check
                     return true;
-                player->ModifyMoney(int32(-RTS1 * GOLD)); // Deducting the money if check passes
+                player->ModifyMoney(-RTS1 * GOLD); // Deducting the money if check passes
                 player->CastSpell(player, 47292); // Level up visual effect to let you know the transaction did occure and money is deducted
                 RemoveAllRacials(player); // Remove all racials
                 player->learnSpell(20599, false); // Diplomacy
@@ -1396,9 +1396,9 @@ public:
 
             case 61: // Night Elf
                 CloseGossipMenuFor(player);
-                if (!player->HasEnoughMoney(int32(RTS1 * GOLD))) // gold check
+                if (!player->HasEnoughMoney(RTS1 * GOLD)) // gold check
                     return true;
-                player->ModifyMoney(int32(-RTS1 * GOLD)); // Deducting the money if check passes
+                player->ModifyMoney(-RTS1 * GOLD); // Deducting the money if check passes
                 player->CastSpell(player, 47292); // Level up visual effect to let you know the transaction did occure and money is deducted
                 RemoveAllRacials(player); // Remove all racials
                 player->learnSpell(20583, false); // Nature Resistance
@@ -1417,9 +1417,9 @@ public:
 
             case 71: // Orc
                 CloseGossipMenuFor(player);
-                if (!player->HasEnoughMoney(int32(RTS1 * GOLD))) // gold check
+                if (!player->HasEnoughMoney(RTS1 * GOLD)) // gold check
                     return true;
-                player->ModifyMoney(int32(-RTS1 * GOLD)); // Deducting the money if check passes
+                player->ModifyMoney(-RTS1 * GOLD); // Deducting the money if check passes
                 player->CastSpell(player, 47292); // Level up visual effect to let you know the transaction did occure and money is deducted
                 RemoveAllRacials(player); // Remove all racials
                 player->learnSpell(20574, false); // Axe Specialization
@@ -1475,9 +1475,9 @@ public:
 
             case 81: // Tauren
                 CloseGossipMenuFor(player);
-                if (!player->HasEnoughMoney(int32(RTS1 * GOLD))) // gold check
+                if (!player->HasEnoughMoney(RTS1 * GOLD)) // gold check
                     return true;
-                player->ModifyMoney(int32(-RTS1 * GOLD)); // Deducting the money if check passes
+                player->ModifyMoney(-RTS1 * GOLD); // Deducting the money if check passes
                 player->CastSpell(player, 47292); // Level up visual effect to let you know the transaction did occure and money is deducted
                 RemoveAllRacials(player); // Remove all racials
                 player->learnSpell(20552, false); // Cultivation
@@ -1495,9 +1495,9 @@ public:
 
             case 91: // Troll
                 CloseGossipMenuFor(player);
-                if (!player->HasEnoughMoney(int32(RTS1 * GOLD))) // gold check
+                if (!player->HasEnoughMoney(RTS1 * GOLD)) // gold check
                     return true;
-                player->ModifyMoney(int32(-RTS1 * GOLD)); // Deducting the money if check passes
+                player->ModifyMoney(-RTS1 * GOLD); // Deducting the money if check passes
                 player->CastSpell(player, 47292); // Level up visual effect to let you know the transaction did occure and money is deducted
                 RemoveAllRacials(player); // Remove all racials
                 player->learnSpell(20557, false); // Beast Slaying
@@ -1517,9 +1517,9 @@ public:
 
             case 101: // Undead
                 CloseGossipMenuFor(player);
-                if (!player->HasEnoughMoney(int32(RTS1 * GOLD))) // gold check
+                if (!player->HasEnoughMoney(RTS1 * GOLD)) // gold check
                     return true;
-                player->ModifyMoney(int32(-RTS1 * GOLD)); // Deducting the money if check passes
+                player->ModifyMoney(-RTS1 * GOLD); // Deducting the money if check passes
                 player->CastSpell(player, 47292); // Level up visual effect to let you know the transaction did occure and money is deducted
                 RemoveAllRacials(player); // Remove all racials
                 player->learnSpell(20577, false); // Cannibalize
